hid version is default for planeta
