--- keyboards/ergohaven/planeta/keymaps/hid/config.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#define VIAL_KEYBOARD_UID {0x3F, 0xB7, 0x5D, 0x50, 0xC1, 0x87, 0xBA, 0x36}
-
-#define VIAL_UNLOCK_COMBO_ROWS { 0, 0 }
-#define VIAL_UNLOCK_COMBO_COLS { 0, 1 }
