--- keyboards/ergohaven/planeta/keymaps/hid/rules.mk
@@ -1,4 +1,0 @@
-VIAL_ENABLE = yes
-RAW_ENABLE = yes
-
-SRC += ./hid_display.c
