--- keyboards/ergohaven/planeta/keymaps/hid/vial.json
@@ -1,29 +1,0 @@
-{
-    "name": "Planeta v2 (HID)",
-    "lighting": "qmk_rgblight",
-    "matrix":{ "rows":5, "cols":12 },
-#include "keyboards/ergohaven/vial_custom_keycodes.json"
-    "layouts":{
-        "keymap":[
-            [{"r":10,"rx":0.6,"ry":0},"0,0",{"y":-0.2},"0,1",{"y":-0.2},"0,2",{"y":-0.2},"0,3",{"y":0.2},"0,4",{"y":0.2},"0,5"],
-            [{"y":0.2},"1,0",{"y":-0.2},"1,1",{"y":-0.2},"1,2",{"y":-0.2},"1,3",{"y":0.2},"1,4",{"y":0.2},"1,5"],
-            [{"y":0.2},"2,0",{"y":-0.2},"2,1",{"y":-0.2},"2,2",{"y":-0.2},"2,3",{"y":0.2},"2,4",{"y":0.2},"2,5"],
-            [{"y":0.2},"3,0",{"y":-0.2},"3,1",{"y":-0.2},"3,2",{"y":-0.2},"3,3",{"y":0.2},"3,4",{"y":0.2},"3,5"],
-            [{"r":0,"x":0.8,"y":0.57},"4,0","4,1"],
-            [{"r":10,"rx":3.6,"ry":4.4},"4,2"],
-            [{"r":20,"rx":4.8,"ry":4.6},"4,3"],
-            [{"r":30,"rx":5.95,"ry":5.05},"4,4"],
-            [{"r":10,"rx":6.13,"ry":3.92},"4,5"],
-
-            [{"r":-10,"rx":8.3,"ry":0.85},"0,6",{"y":-0.2},"0,7",{"y":-0.2},"0,8",{"y":0.2},"0,9",{"y":0.2},"0,10",{"y":0.2},"0,11"],
-            [{"y":-0.2},"1,6",{"y":-0.2},"1,7",{"y":-0.2},"1,8",{"y":0.2},"1,9",{"y":0.2},"1,10",{"y":0.2},"1,11"],
-            [{"y":-0.2},"2,6",{"y":-0.2},"2,7",{"y":-0.2},"2,8",{"y":0.2},"2,9",{"y":0.2},"2,10",{"y":0.2},"2,11"],
-            [{"y":-0.2},"3,6",{"y":-0.2},"3,7",{"y":-0.2},"3,8",{"y":0.2},"3,9",{"y":0.2},"3,10",{"y":0.2},"3,11"],
-            [{"r":-10,"rx":7.72,"ry":4.1},"4,6"],
-            [{"r":-30,"rx":8.05,"ry":5.55},"4,7"],
-            [{"r":-20,"rx":9.12,"ry":4.95},"4,8"],
-            [{"r":-10,"rx":10.3,"ry":4.58},"4,9"],
-            [{"r":0,"x":1.2,"y":-1.18},"4,10","4,11"]
-        ]
-    }
-}
