+++ keyboards/ergohaven/planeta/keymaps/v2/keymap.c
@@ -1,5 +1,13 @@
 #include QMK_KEYBOARD_H
 #include "ergohaven.h"
+#include "hid_display.h"
+#include "raw_hid.h"
+#include "rev2.h"
+
+// #ifdef AUDIO_ENABLE
+// float base_sound[][2] = SONG(TERMINAL_SOUND);
+// float caps_sound[][2] = SONG(CAPS_LOCK_ON_SOUND);
+// #endif
 
 const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
         [_BASE] = LAYOUT( \
@@ -34,3 +42,55 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
                  _______, _______, _______, _______, _______, _______,         _______, _______, _______, _______, _______, _______
        ),
 };
+
+
+
+/* Active Layer processing */
+// layer_state_t layer_state_set_keymap(layer_state_t state) {
+//     if (is_display_enabled()) {
+//         display_process_layer_state(get_highest_layer(state));
+//     }
+
+//     return state;
+// }
+
+/* Raw HID processing*/
+void raw_hid_receive_kb(uint8_t *data, uint8_t length) {
+    dprintf("raw_hid_receive - received %u bytes \n", length);
+
+    // if (is_display_enabled()) {
+        display_process_raw_hid_data(data, length);
+    // } else if (is_keyboard_master() && !is_display_side()) {
+        // dprint("RPC_ID_USER_HID_SYNC \n");
+        // transaction_rpc_send(RPC_ID_USER_HID_SYNC, length, data);
+    // }
+}
+
+void hid_sync(uint8_t initiator2target_buffer_size, const void *initiator2target_buffer, uint8_t target2initiator_buffer_size, void *target2initiator_buffer) {
+    if (is_display_enabled()) {
+        display_process_raw_hid_data((uint8_t *)initiator2target_buffer, initiator2target_buffer_size);
+    }
+}
+
+// void keyboard_post_init_user() {
+    // sync received hid data
+    // transaction_register_rpc(RPC_ID_USER_HID_SYNC, hid_sync);
+// }
+
+layer_state_t layer_state_set_user(layer_state_t state) {
+    if (is_display_enabled()) {
+        display_process_layer_state(get_highest_layer(state));
+    }
+    //   #if defined(AUDIO_ENABLE)
+    //     static bool is_base_on = false;
+    // if (layer_state_cmp(state, _BASE) != is_base_on) {
+    //         is_base_on = layer_state_cmp(state, _BASE);
+    //         if (is_base_on) {
+    //             stop_all_notes();
+    //         } else {
+    //             PLAY_SONG(base_sound);
+    //         }
+    //     }
+    // #endif
+    return state;
+}
