+++ keyboards/ergohaven/planeta/keymaps/v2/rules.mk
@@ -1,1 +1,4 @@
 VIAL_ENABLE = yes
+RAW_ENABLE = yes
+
+SRC += ./hid_display.c
