+++ keyboards/ergohaven/planeta/keymaps/v2/vial.json
@@ -1,31 +1,29 @@
 {
-    "name": "Planeta v2",
+    "name": "Planeta v2 (HID)",
     "lighting": "qmk_rgblight",
     "matrix":{ "rows":5, "cols":12 },
 #include "keyboards/ergohaven/vial_custom_keycodes.json"
     "layouts":{
         "keymap":[
-            [{"c":"#aaaaaa","x":3},"0,3",{"x":7.25},"0,8"],
-            [{"x":2,"y":-0.85},"0,2",{"x":1},"0,4",{"x":5.25},"0,7",{"x":1},"0,9"],
-            [{"x":1,"y":-0.8499999999999999},"0,1",{"x":3},"0,5",{"x":3.25},"0,6",{"x":3},"0,10"],
-            [{"c":"#777777","y":-0.8},"0,0",{"c":"#aaaaaa","x":13.25},"0,11"],
-            [{"c":"#cccccc","x":3,"y":-0.5},"1,3",{"x":7.25},"1,8"],
-            [{"x":2,"y":-0.8500000000000001},"1,2",{"x":1},"1,4",{"x":5.25},"1,7",{"x":1},"1,9"],
-            [{"x":1,"y":-0.8499999999999999},"1,1",{"x":3},"1,5",{"x":3.25},"1,6",{"x":3},"1,10"],
-            [{"y":-0.7999999999999998},"1,0",{"x":13.25},"1,11"],
-            [{"x":3,"y":-0.5},"2,3",{"x":7.25},"2,8"],
-            [{"x":2,"y":-0.8500000000000001},"2,2",{"x":1},"2,4",{"x":5.25},"2,7",{"x":1},"2,9"],
-            [{"x":1,"y":-0.8500000000000001},"2,1",{"x":3},"2,5",{"x":3.25},"2,6",{"x":3},"2,10"],
-            [{"c":"#aaaaaa","y":-0.7999999999999998},"2,0",{"c":"#cccccc","x":13.25},"2,11"],
-            [{"x":3,"y":-0.5},"3,3",{"x":7.25},"3,8"],
-            [{"x":2,"y":-0.8500000000000001},"3,2",{"x":1},"3,4",{"x":5.25},"3,7",{"x":1},"3,9"],
-            [{"x":1,"y":-0.8500000000000005},"3,1",{"x":3},"3,5",{"x":3.25},"3,6",{"x":3},"3,10"],
-            [{"c":"#aaaaaa","y":-0.7999999999999998},"3,0",{"c":"#cccccc","x":5},"4,5",{"x":1.25},"4,6",{"x":5},"3,11"],
-            [{"x":3,"y":-0.5},"4,1",{"x":7.25},"4,10"],
-            [{"x":2,"y":-0.8499999999999996},"4,0",{"x":9.25},"4,11"],
-            [{"x":4,"y":-0.9000000000000004},"4,2",{"x":5.25},"4,9"],
-            [{"x":5,"y":-0.8499999999999996},"4,3",{"x":3.25},"4,8"],
-            [{"x":6,"y":-0.8000000000000007},"4,4",{"x":1.25},"4,7"]
+            [{"r":10,"rx":0.6,"ry":0},"0,0",{"y":-0.2},"0,1",{"y":-0.2},"0,2",{"y":-0.2},"0,3",{"y":0.2},"0,4",{"y":0.2},"0,5"],
+            [{"y":0.2},"1,0",{"y":-0.2},"1,1",{"y":-0.2},"1,2",{"y":-0.2},"1,3",{"y":0.2},"1,4",{"y":0.2},"1,5"],
+            [{"y":0.2},"2,0",{"y":-0.2},"2,1",{"y":-0.2},"2,2",{"y":-0.2},"2,3",{"y":0.2},"2,4",{"y":0.2},"2,5"],
+            [{"y":0.2},"3,0",{"y":-0.2},"3,1",{"y":-0.2},"3,2",{"y":-0.2},"3,3",{"y":0.2},"3,4",{"y":0.2},"3,5"],
+            [{"r":0,"x":0.8,"y":0.57},"4,0","4,1"],
+            [{"r":10,"rx":3.6,"ry":4.4},"4,2"],
+            [{"r":20,"rx":4.8,"ry":4.6},"4,3"],
+            [{"r":30,"rx":5.95,"ry":5.05},"4,4"],
+            [{"r":10,"rx":6.13,"ry":3.92},"4,5"],
+
+            [{"r":-10,"rx":8.3,"ry":0.85},"0,6",{"y":-0.2},"0,7",{"y":-0.2},"0,8",{"y":0.2},"0,9",{"y":0.2},"0,10",{"y":0.2},"0,11"],
+            [{"y":-0.2},"1,6",{"y":-0.2},"1,7",{"y":-0.2},"1,8",{"y":0.2},"1,9",{"y":0.2},"1,10",{"y":0.2},"1,11"],
+            [{"y":-0.2},"2,6",{"y":-0.2},"2,7",{"y":-0.2},"2,8",{"y":0.2},"2,9",{"y":0.2},"2,10",{"y":0.2},"2,11"],
+            [{"y":-0.2},"3,6",{"y":-0.2},"3,7",{"y":-0.2},"3,8",{"y":0.2},"3,9",{"y":0.2},"3,10",{"y":0.2},"3,11"],
+            [{"r":-10,"rx":7.72,"ry":4.1},"4,6"],
+            [{"r":-30,"rx":8.05,"ry":5.55},"4,7"],
+            [{"r":-20,"rx":9.12,"ry":4.95},"4,8"],
+            [{"r":-10,"rx":10.3,"ry":4.58},"4,9"],
+            [{"r":0,"x":1.2,"y":-1.18},"4,10","4,11"]
         ]
     }
 }
